Do more cleanups
We keep cleaning up some files (e.g. lock files) in the git repos (main
repo & submodules) on build start, but in addition we do this also on
build end.

In a similar fashion, if git config cleanup is enabled, we also clean
out git configs at the beginning and end of a build.