Move commands configOptions to configfile package
configOptions has grown too large, encompassing too many
responsibilities.

This change separates the "ConfigFile" command argument from the
config file handling logic, in addition to moving some finder functions
to common.Config where they make more sense to be.

In addition, using the "configfile" package to load a config now takes
a mutation argument, putting a lock around modifications to the config,
removing a race between opening, modification, and saving.

This fixes an unknown bug where the command line argument --user can be
persisted to the config, in the event that the config is saved due to
a token refresh.

With a clear path of where mutations to the config can occur, we no
longer need to shallow copy the common.Config on each fetch.