Allow OS overwrite via ShellScriptInfo
The runner manager can run on a different OS than the job, e.g. when
the k8s executor is used, the runner manager runs on a linux node, while
the jobs are set up to run on windows nodes.
This adds the option to overwrite that via the ShellScriptInfo, by
setting its `RuntimeOS`.

We use this for building up the git credential helper command, which
needs to be slightly different on different combinations of OS and
powershell/pwsh.