Make submodule `--remote` more resilient
When users
- "auto-bump" submodules with `--remote`
- and in the superproject a specific branch is set in `.gitmodules`
- and the default branch in the submodule repo is not that branch

then updating the submodules might fail with

> fatal: Unable to find refs/remotes/origin/<submoduleBranch> revision in submodule path '<submoduleDir>'

With this change, on the 2nd attempt we try to fetch the remote's heads,
which makes the ref available locally.

We only do this on the retry and when the `--remote` flag is actually
used, as a fall-back, so that we do not (unnecessarily) pull in
potentially a lot of remote heads.

Another issue was observed:

> fatal: Needed a single revision
> Unable to find current origin/master revision in submodule path '<submoduleDir>'

This seems to stem from old git versions, which have `origin/master`
hardcoded, instead of considering the remote's default branch.
For this we don't do anything, except we skip the tests to "auto-bump"
without a branch explicitly set in `.gitmodules`.