shells: Guard git-clone --revision with version check
Because only Git from version 2.49 supports option `--revision` for
`git-clone(1)`, add a if guard around the use of that option. This if
guard tests if the Git version is 2.49 or above. It uses option
`--version-sort` for `sort(1)`.

At the moment this if guard is only implemented in the `BashWriter`,
other shell do not yet support this version check. Because the majority
of runners uses bash anyway, we'll take this as a first iteration.