Fix cache extractor redownloading up-to-date caches for Go Cloud URLs
For pre-signed URLs, the runner helper checks that the `Last-Modified`
header of the cache blob is later than the modification time of the
downloaded cache file. This avoids having to download an unchanged
cache file.

This was previously not happening for Go Cloud URLs. This commit
adds support for this checking.

Note that the Last-Modified time appears to be updated automatically
by Go Cloud, so we don't to set this header as we do with presigned
URLs in the cache archiver.

Changelog: fixed