Pin zstandard version and specify checksum
A new version (1.5.7) of zstandard was release Feb 26. The checksums
specific in the package are wrong. It looks like the upstream package
was changed (and presumably re-released without a version change) AFTER
the chocolatey package was created, which changed the actual checksum.

See https://github.com/facebook/zstd/issues/4326.

This commit both specifies the correct checksum to fix the immediate
problem, and pins the zstandard version so this doesn't happen in the
future.