Trigger the downstream pipeline to test OS packages
Trigger a downstream pipeline to test OS packages.

The is done via a s-stage approach, where a first job generates an
artifact with the test parameters for the downstream pipeline, and a
second job triggers the downstream pipeline. We must do it this way because:
1. We want this pipeline to fail if the downstream pipeline fails, so we
   must use the `strategy:depends`. With this, GitLab (not runner)
   triggers the pipeline.
2. Some of the test parameters (the to/from versions) can only be
   computed by runner in the runner repo.

The 2-stage approach is the easiest way to do this.

The tests are implemented in
https://gitlab.com/avonbertoldi/gitlab-runner-package-tests. The test
suite is currently while quite minimal, but I'll add to it in the near
future. I'll also move the project to a more permanent place, likely in
gitlab-org/ci-cd/runner-tools.

Fixes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38505