Fix runner_wrapper gRPC API client
We've decided we want to support both TCP and UNIX sockets
for the wrapper gRPC server. UNIX sockets are referenced
using the `unix:///path/to/socket` URI.

For consistency, TCP sockets were supposed to be refrenced
using the `tcp://address:port` URI. This, however, doesn't work
with the gRPC library. Such address is handled as `address:port:443`.
For some reason gRPC library doesn't parse that URI, assumes the
`address:port` to be just the address part, with no port specified
and adds a default `:443` port. Which, obviously, causes connection
errors.

If the address were specified as just `address:port`, then it would
work. But that makes the TCP and UNIX socket definitions inconsistent
in the client and may confuse users.

This commit brings an update that allows us to keep the `tcp://`
part, while transforms the TCP address to a form expected by the
gRPC library.