Exclute helpers/runner_wrapper/api/v* tags from version evaluation
When compiling Runner `pre` versions from the `main` branch,
we use the `ci/version` script to establish the version number.
Among all things it does, it searches for the last commit and
calculates the number of commits that happened since then.

With this, the `x.y.z~pre.N.gSHA` version has an always growing
value and can be used by users and tools (like APT for Debian
packages) to decide which one is the latest.

With the runner_wrapper API submodule added recently with
2946f23c4d0ec98a8d80d45a536e59b45c153863, we've started creating
additional tags to version the submodule. And that's now affecting
the Runner versioning system.

This commit excludes `helpers/runner_wrapper/api/v*` tags from
the version number calculation, as version management of the Wrapper
API submodule is a totally independent thing and should not be
collide with Runner versioning itself.