Fix incept tests
Some incept test are failing with:

    "This job could not start because it could not retrieve the needed artifacts."

This is because some jobs in the downstream pipeline depend on the
"prebuilt-helper-images" from this pipeline, but the triggering job
itself did not depend on that job. So the downstream pipeline was
triggered before it's dependencies were met.

Fix that, and also document the dependencies.