Fix usage log timestamp generation
While working on the usage log
(https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/5202/)
and preparing tests for it, timestamp generation was refactored
in a way that allows the time to be stubbed during tests.

Unfortunately, it was done in a very wrong and stupid way, that now
saves the time of the first usage log entry creation since when runner
was started, and then uses that timestamp forever...

This commit fixes this bug and ensures that we have a test coverage for
this specific problem.