Split up client.doJSONWithPAT
The call tree is all backwards with doJSON and doJSONWithPAT. The more
generic doJSON calls the more specific doJSONWithPAT, passing empty
arguments which the latter has to handle.

Reverse the call tree, with doJSONWithPAT only doing the PAT specific
things, and then calling doJSON which handles everything else.

This is the first step to completely eliminating everything
*JSONWithPAT*.