Use Runner-Token instead of Private-Token where applicable
In these API calls, the token being used is, AFAIU, the runner
registration token (it's certainly not a job-token or private-token). In
coordination with the Cell's folks, we've decided to add a new
Runner-Token header for these API calls to make it clear what token is
being used. The Cells folks will update the HTTP router to accommodate
this at a latter time.

See discussion starting here https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/5470#note_2422763759