Log failure to remove a container
The original report
https://gitlab.com/gitlab-com/request-for-help/-/issues/2600 confused me
because it mentioned the error

    "Failed to cleanup volumes" error

I'd also expected to see a "failed to remove container" error, but their
was none, and docker won't let you remove a volume that is attached to a
running container, so the container has to be removed first.

Anyway, it's generally not a good idea to silently swallow errors.