Expose started_at and finished_at values in Usage Log job details
Runner usage analysis may require information about when Runner started
handling a particular job and when it finished handling it. For example,
some billing strategies may allocate the generated usage and consumed
duration to a given month given the timestamp of when the job was
started or finished.

Therefore, this commit exposes these two values explicitly as part of
job details.

The existing `timestamp` field is still left and it represents the exact
moment when the usage log record was generated, which may be few
milliseconds later than when the job was marked as finished and
therefore what the `finished_at` will expose.

Closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38709