Update step-runner to v0.8.0
One API in step-runner changes in 0.8.0. (api.ListenSocketPath =>
api.DefaultSocketPath), so some references and to be updated.

This also triggered update of a number of docker modules several majour
versions. This broke all manner of code which I'll fix in the next
commit.

Finally, this also bumped the min go version to `1.23.6`, which is
actually want were using in runner anyway, so that's fine.