Refactor git cred helper setup
The setup of the git credential helper has been pushed to the shell
writer. With that we can react on differences between shells and
versions thereof at runtime.

We expose this through one method on the shell writer, handling all the
business of:
- replacing / pruning all other cred helpers
- ensuring the git commands that run have their arguments escaped
  properly
- configuring both the helper itself and the username

What we do for the different versions:
- Starting from 7.3 we can pass on args without special handling
- Before 7.3 we generally need to handle args explicitly, ie. escape
  "special" things
- For 7.2.x it depends on `PSNativeCommandArgumentPassing`¹:
  - if this feature is disabled, we also need to handle args specially
  - if this feature is enabled, we need to use it, but have control over
    exact behaviour via `PSNativeCommandArgumentPassing`

To not run into issues with `$PSNativeCommandArgumentPassing`, we run
all that in its own script block: this preference variable should only
take effect for the current and child scopes, thus only for the (git)
commands we run inside that script block.

¹ https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_preference_variables?view=powershell-7.5#psnativecommandargumentpassing