Prepare failures should be RunnerSystemFailure not ScriptFailure
Unless a valid `JobFailureReason` is specified by wrapping an `error` in
a `BuildError`, errors will eventually be wrapped to a `BuildError` with
a `JobFailureReason` of `ScriptFailure`. However, errors that occur in
the `Prepare` and post `after_script` (i.e. cache/artifact upload)
phases should not map to `ScriptFailures` since we're not even running a
user script! RunnerSystemFailure is the best alternative
JobFailureReason.