Fix final job duration handling
With https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/5484
we've introduced a new way to track final job duration. That allowed
to track this value properly with the new usage logs, but also affected
the existing logging and metrics mechanisms.

Unfortunately, this Merge Request introduced a bug that technically have
erased the duration value in cases where job succeeds or is failed - so
majority of them. The duration is tracked properly only when there is a
runner interrupt signal received or when the job is canceled. Which in
case of normal operations doesn't happen often.

This change fixes that bug.