Inject Context into environmentDialContext
To fix https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38725, we'll
need to inject a context.Context into environmentDialContext, which is
called by connectDocker. Let's add a Context argument to connectDocker
and pipe it down to environmentDialContext.

This change doesn't actually fix anything, but I'll use later to help
fix the bug.