Add queue_size and queue_depth metrics
As described in
https://gitlab.com/gitlab-org/ci-cd/shared-runners/infrastructure/-/issues/298+,
insight into what is the size of the jobs queue that potentially may be
executed on a given runner is very limited. Not only it's very
generalized and partitioned by three runner types, it's also available
only for GitLab instance owners.

With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/192572+ we
will get information about the queue size and queue search depth in the
job payload.

Adds support for these values and exports them as a runner metric,
making the information available for every runner owner.