Integration tests for docker+autoscaler timed out job
This test runs a job that times out, and subsequently ensures the
containers and network are deleted.

Unfortunately there isn't an API in our official docker client to list
volumes (to ensure they also are deleted), and I couldn't justify adding
a production API to be used only in one test. I did test this manually
though to make sure it works.