Change metadata handling for local caches
This changes the file name of the metadata file locally, the cache dir
now looks something like:
```
<cacheDir>
└── <group>
    └── <project>
        ├── d0048ab4d378ceb85c26ae020bcfc602b6fa1fe02db339a25dc79b77dd991af4  # for hashed cache keys
        │   ├── cache.zip
        │   └── metadata.json
        └── some-random-cache-key                                             # for human readable cache keys
            ├── cache.zip
            └── metadata.json
```

Also, we pass on metadata to the cache-archive regardless of upload
mechanism (GoCloud vs. pre-signed URL), so that for any of these the
local cache metadata file is created correctly.

The metadata key for the used cache key has been changed slightly, to
make it consistent across the board: when attaching metadata via http
headers, the header might be canonicalized and therefore slightly
diverge between local and cloud metadata.