Add dockerTunnel type
This object and constructor encapsulates creation of an
executors.Client, and the associated docker options and credentials
necessary to subsequently create a docker.Client that tunnels docker
commands through said executors.Client.

The code in the dockerTunnelFactory function is taken verbatim (perhaps
with some minor cleanup) from the connectDocker function.