Add dockerConnection type
This type encapsulates docker.Client and executors.Client instances, and
treats them as one. Because it embeds the docker.Client instance, it can
transparently replace references to the executor's docker.Client
instance. It also holds a reference to the cancel function to the
log-lived context used to tunnel to remote docker daemons.

The code in the newDockerConnecton function is similarly taken verbatim from the connectDocker function.

I'll use this and the previous type in the next commit to reimplement connectDocker.