Ensure BuildErrors have FailureReason
Over in
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/5531 we
tried to improve the granularity and correctness of job failure reasons.
Specifically, in 524e0365a5d06dad143ae77278c4e20ccb15c375, with
asBuildError,  we attempted to make failures in `prepare` and
`after_script` stages have a failure reason of `runner_system_failure`.
However, we missed the case where the called code returns a BuildError
WITHOUT a FailureReason! Such a BuildError will be interpreted
as/converted to a script_failure downstream, which is what we were
trying to prevent.

Let's handle that case explicitly in asBuildError!