Set helper_image_flavor to "ubi-fips" when fips mode is enabled
https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38273 describes a
situation where jobs fail when runner is fips enabled and running in
so-called "air-gapped" environments. The failure occurs because runner
tries to pull the helper image, but the pull fails because there's no
access to the internet.

The thing is, the ubi-fips-flavoured image already exists in the FS as
an exported image, and we want to (are meant to) use that helper image!

All we have to do for this to happen is to set the `helper_image_flavor`
top `ubi-fips`, and runner will do the right thing.

Initially I though to set this when running the install command, but
this is an executor-specific setting, and at that time there is are
configured runners. The right time to do this is on runner registration.