Do not sanitize hashed cache keys
When we hash cache keys, we don't have to do the sanitation. For this,
the sanitation has been pulled up into newCacheConfig func, to have one
place where we handle that.

Until we drop support for not hashing cache keys, we still need to keep
track of what an unhashed cache key would have looked like, so that we
can "upgrade" unhashed local cache files. This is also handled in the
newCacheConfig func, and is exposed on the cacheConfig struct.