Bridge releases with Hosted Runners
To make it easier for Runners Platform team to manage regular updates of
Hosted Runners, this job is added to start noting properly released Pre
and Stable versions.

The job is triggered at the end of stable tags and `main` branch
pipelines, in the canonical and security forks. Jobs from security forks
are manualy.

In the future, we may consider turning that into a trigger, that will
automatically update some of the runners.