[k8s] Allow to use helper image from this pipeline
withDevHelperImage searches the artifact from a previous job
("(development|bleeding|stable) docker image") for a helper image we can
use for tests, and sets it on the build. With that, we can run the
helper image (and the helper binary) from this very change, and not only
the latest released one.

We need to ensure those images have been pushed to a registry kubernetes
can consume it from. Therefore we set up explicit dependencies of the
kubernetes integration test jobs on the docker images jobs. Thus the
kubernetes integration test jobs now run later in the pipeline.

To not prolong the pipeline runtime too much, the dependency between the
"(helper|runner) images" and the "(development|bleeding|stable) docker
images" jobs has been changed from "dependencies" to "needs". This lets
the kubernetes integration test and all dependant jobs run as soon as
possible, not waiting for the previous stages to have finish. A
side-effect of that is, that we potentially publish more images,
potentially some where other tests have not succeeded yet.

This all is opt-in, by default the build's helper image won't be
set/overwritten.