Set ALPINE_DEFAULT_VERSION to latest
Why this? Why now? To take advantage of the performance optimizations
available in git since version 2.49,0, which is available in alpine 3.22
and later. Today our `latest` points to 3.22.

See https://gitlab.com/gitlab-com/gl-infra/production-engineering/-/issues/26837

Note that we need the leading `-` because there's no delimiter between
the numbered versions.