Pass cache metadata via env
When calling the cache archiver, we pass the cache archive metadata via
the environment rather than CLI args. With that, calling the cache
archiver is backwards compatible: The call will also succeed for older
versions of the helper binary (e.g. if users have pinned the
helper_image to a specific version).

When users run with an older version of the helper binary /
cache_archiver, hashed cache keys would still work, but metadata
alongside the local cache artifact or on the blob won't be managed.

To make passing on metadata safe and to not have to care about escaping
and whatnot, the metadata flag's type has been switched to using plain
JSON format.