Get a real unique name for a build
For some things we want to have a real unique name per build && runner.
Other approaches use the shortened token to generate unique names, but
that can lead to collisions when tokens are similar enough. This gives
us a real unique name, which hashes a couple of things from the
build.