Add timeouts to all docker-machine command executions
In https://gitlab.com/gitlab-com/gl-infra/production-engineering/-/issues/27396
we saw that `docker-machine create` hung for several days, causing
pipelines to be stuck because the maximum number of VMs were in the
`creating` state. While we don't know why the command hung, this commit
adds a timeout on every `docker-machine` command to prevent this from
happening.

The timeouts are very conservative with a 1-hour limit. The Stop
timeout remains at 1 minute.