Add support for signing and notarizing macOS binaries
This commit updates `scripts/sign-binaries` to handle the signing and
notarizing of macOS binaries with Google Cloud HSM. This is all done on
a Linux system without needing macOS.

Signing macOS binaries requires `rcodesign` with PKCS11 support to work
with the HSM. The private key never leaves the HSM during the signing
process.

Notarization uses the approach used in
https://gitlab.com/gitlab-org/rust/knowledge-graph/-/merge_requests/117
where the API key is stored in the `APPSTORE_CONNECT_API_KEY_FILE` CI/CD
variable. As notarization can take some time to complete, the script
currently submits the request for notarization but does not wait for
successful notarization because we don't want to block the entire
release process if that fails.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/3544