Fix flaky Docker integration tests
Previously `Test_CaptureServiceLogs` failed in some of the race unit
tests because the log output of PostgreSQL and Redis services were
interleaved on multiple lines. `InlineServiceLogWriter` makes multiple
`Write()` calls to the same sink, so the output in
https://gitlab.com/gitlab-org/gitlab-runner/-/jobs/11114656794 had lines
such as:

```
[service:db-postgres] [service:redis-cache] 2025-08-23T04:36:09.130700918Z
```

Fix this by testing for the prefixes and expected service strings
separately to avoid having to match across multiple lines.