Fix arch label for IBM PPC arch
Over in 3a3f0ec83501f9741c193dac1a249440bda47b0d we reimplemented
package building in mage instead of scrips and makefile targets. The
following snippet from the diff of that commit shows that for rpmxppc,
the package-arch should be ppc64le:

    diff --git a/Makefile.package.mk b/Makefile.package.mk
    deleted file mode 100644
    index a03bc4b28..000000000
    --- a/Makefile.package.mk
    +++ /dev/null
    ...
    -.PHONY: package-deb-ibm
    -package-deb-ibm: package-deps package-prepare
    -   $(MAKE) package-deb-arch ARCH=s390x PACKAGE_ARCH=s390x
    -   $(MAKE) package-deb-arch ARCH=ppc64le PACKAGE_ARCH=ppc64el
    ...
    -.PHONY: package-rpm-ibm
    -package-rpm-ibm: package-deps package-prepare
    -   $(MAKE) package-rpm-arch ARCH=s390x PACKAGE_ARCH=s390x
    -   $(MAKE) package-rpm-arch ARCH=ppc64le PACKAGE_ARCH=ppc64le

But in the new mage version it was set to ppc64el. This commit restores
it to ppc64le.