Remove duplicate prefix in docker service containers
This change set removes the duplicate prefixes in the docker service
containers that have been introduced in
https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/4913.

The duplicate prefixes have been added in
`createServiceContainerConfig`. The prefixes already existed for the
default labels and `otherLabels` in `prepareContainerLabels`, but not
yet for the labels from the configuration for the docker containers.

This patch avoids having to prefix the labels in
`createServiceContainerConfig` and leaves it up to
`prepareContainerLabels` entirely.

Closes https://gitlab.com/gitlab-org/gitlab-runner/-/issues/38674