Rename Connect() to TerminalConnect()
Some executors have a Connect() function, which means they implement
the InterativeTerminal interface.

It's likely that we'll use the name Connect() for connecting to the
build environment via CI Functions, which will also likely replace
the interative terminal functionality in the future.

To avoid a naming collision, we rename Connect() to TerminalConnect().