Add correlation_id to "Update job..." log line
Job updating API call is the only that logs the attempt to make
a call before it's done and the response is received. And today,
despite the fact that runner does generate the correlation ID,
that's a log entry that's part of the API flow but misses the
correlation ID.

This commit adds also the `correlationIDLogField` const to make sure
that correlation_id field is defined with consistent naming across
different log entries that may want to use it.