Use passed context in NewStepsDocker.Exec()
As far as I've been able to determine, the only difference between the context embedded in the executor
and the one passed to startAndWatchContainer is that the latter wraps
the former in a cancellable context. I don't see a good reason to use
the former over the latter, and using a single one keeps things clearer.
Like, I wrote this code and am now confused why we use the executors
context in that one place.