Update cache S3 SSE Key ID docs
When the runner runs in account1 and its cache is stored in account2,
specifying the SSE ID/alias results in searching for the key in
account1 instead of account2 where it actually exists.
This can be achieved by specifying the ARN of the key which has
the account information embedded in itself.