Do not fail install if gitlab-runner service commands not available
`gitlab-runner status` uses https://github.com/kardianos/service and
attempts to find the right command to manage the GitLab Runner service.
On Linux systems, it falls back to running `service` from `initscripts`
if `systemctl` or other service managers are not detected. However, in
some environments, such as Fedora 43 containers, `initscripts` are not
installed by default. As a result, the installation of the Runner fails
in the post-install script.

To avoid this install failure, just skip the service management part of
the post-install and pre-remove scripts if `gitlab-runner status`
fails with `executable file not found`.

Relates to https://gitlab.com/gitlab-org/gitlab-runner/-/issues/39006