Include go.mod changes when checking modules in pipeline
This change set ensures that we also check the integrity of the `go.mod`
file in addition to the `go.sum` file after running `go mod tidy` in the
pipeline. The `go.mod` might also be updated with a call to `go mod
tidy` - like changing modules from being used directly / indirectly.